M3-157 Refactor : countCurrentPixelByUserId 를 Jpql 로 변경

+++ src/main/java/com/m3pro/groundflip/domain/dto/pixel/PixelCountResponse.java
@@ -13,7 +13,7 @@ import lombok.NoArgsConstructor;
 @Schema(title = "픽셀 개수")
 public class PixelCountResponse {
 	@Schema(description = "현재 차지하고 있는 픽셀 개수", example = "5")
-	private Integer currentPixelCount;
+	private Long currentPixelCount;
 
 	@Schema(description = "지금까지 차지한 픽셀 개수", example = "5")
 	private Long accumulatePixelCount;
